[#16] Refactor: 설정과 관련된 클래스 작성
